+++ _includes/service-worker-css-blog-post-urls.js
@@ -1,1 +1,1 @@
-"/assets/styles/style.blog.post.css?rev=efce2820913895675358eb649330c739",
+"/assets/styles/style.blog.post.css?rev=6684f33c34b6f18f6b700cdf4f6ad94f",