+++ _jsbuild/index.blog.js
@@ -1,6 +1,7 @@
 /*  */
 import { cookieConsent } from './common/cookie-consent.js'
 import { lazyLoadImages } from './common/lazy-load-images'
+import { disqus } from './blog/disqus'
 
 document.addEventListener('DOMContentLoaded', () => {
   lazyLoadImages('.blog-image')
@@ -8,4 +9,5 @@ document.addEventListener('DOMContentLoaded', () => {
 
 window.addEventListener('load', () => {
   cookieConsent()
+  disqus()
 })